+++ DFM.Core/Entities/Bases/ISummarizable.cs
@@ -6,7 +6,7 @@ namespace DFM.Core.Entities.Bases
     public interface ISummarizable : IEntity
     {
         IList<Summary> SummaryList { get; set; }
-        Int32 Time { get; set; }
+        Int16 Time { get; set; }
 
         Double CheckUpIn(Category category);
         Double CheckUpOut(Category category);
