+++ DFM.Core/Entities/Month.cs
@@ -19,7 +19,7 @@ namespace DFM.Core.Entities
 
         public virtual Int32 ID { get; set; }
 
-        public virtual Int32 Time { get; set; }
+        public virtual Int16 Time { get; set; }
 
         public virtual Year Year { get; set; }
