+++ DFM.Core/Entities/Schedule.cs
@@ -19,7 +19,7 @@ namespace DFM.Core.Entities
 
         public virtual Int32 ID { get; set; }
 
-        public virtual Int32 Times { get; set; }
+        public virtual Int16 Times { get; set; }
         public virtual Boolean Boundless { get; set; }
 
         public virtual DateTime Begin { get; set; }
